refactor(compaction): extract resolveTokenThreshold as independent helper
Move the inline percentage-parsing function from agent-runner-memory.ts
into src/config/token-threshold.ts for reuse.

Fixes #87136
